Check for existence of directories in run project script

+++ run_project.sh
@@ -4,5 +4,10 @@
 quarto render
 
 # move the products over to the public_html folder after removing the old ones
-rm -r ~/public_html/research/multiracial_sim
-cp -r _products ~/public_html/research/multiracial_sim
+if [ -d ~/public_html/research/multiracial_sim ]; then
+  rm -r ~/public_html/research/multiracial_sim
+fi
+
+if [ -d _products ]; then
+  cp -r _products ~/public_html/research/multiracial_sim
+fi
